requirements.txt
openai
watchdog
pymongo
python-dotenv
pyobjc-framework-Vision
pyobjc-framework-Quartz
pyobjc-framework-Cocoa
pyobjc-framework-AVFoundation
pyobjc-framework-Speech
pillow
typer
rich
SpeechRecognition
pydub
